Fix FileIONative on linux. Migrate to Kotlin Gradle DSL. Add CI.

+++ utils/src/nativeMain/kotlin/sample/utils/FileIONative.kt
@@ -8,10 +8,10 @@ actual fun readAllBytes(path: String): ByteArray {
     val file = fopen(path, "rb") ?: throw Exception("File not found.")
     try {
         fseek(file, 0, SEEK_END)
-        val fileLength = ftell(file)
+        val fileLength = ftell(file) // Return Long on linux but Int on mingw.
         fseek(file, 0, SEEK_SET)
 
-        val buffer = ByteArray(fileLength)
+        val buffer = ByteArray(fileLength.toInt())
         buffer.usePinned {
             fread(it.addressOf(0), 1u, fileLength.toULong(), file)
         }
